requirements.txt
# pip install -r requirements.txt

# Twitter Scraping --------------------------------------------
tweety-ns
httpx[http2]
emoji


# Database --------------------------------------------
pandas

# Data analysis --------------------------------------------
networkx
matplotlib
plotly

# User flagging --------------
git+https://github.com/casics/nostril.git    #for randomly generated string detection


# Web UI --------------------------------------------
fastapi


# Misc --------------------------------------------
tqdm
